#made edits for pr 8

+++ server/src/routes/api/v1/reviewsRouter.js
@@ -5,11 +5,16 @@ const reviewsRouter = new express.Router();
 
 reviewsRouter.delete("/:id", async (req, res) => {
   const id = req.params.id;
-  try {
-    await Review.query().deleteById(id);
-    res.status(200).json({ message: "Review successfully deleted" });
-  } catch (err) {
-    res.status(500).json({ error: err });
+  const userId = req.user.id;
+  const reviewUserId = req.body.reviewUserId;
+
+  if (userId === reviewUserId) {
+    try {
+      await Review.query().deleteById(id);
+      res.status(200).json({ message: "Review successfully deleted" });
+    } catch (err) {
+      res.status(500).json({ error: err });
+    }
   }
 });
